third_party/blink/web_tests/wpt_internal/attribution-reporting/trigger-data-sanitization.sub.https.html

<!doctype html>
<meta charset=utf-8>
<meta name=timeout content=long>
<meta name=variant content="?source-type=event">
<meta name=variant content="?source-type=navigation">
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>
const sourceTypes = {
  event: {
    method: 'img',
    cardinality: 2,
  },
  navigation: {
    method: 'open',
    cardinality: 8,
  },
};

const sourceType = (new URLSearchParams(location.search)).get('source-type');

const {method, cardinality} = sourceTypes[sourceType];

const trigger_data_cardinality_promise_test = value =>
  attribution_reporting_promise_test(async t => {
    await registerAttributionSrc(t, {
      source: {
        source_event_id: '999',
        destination: `https://{{host}}`,
      },
      method,
    });

    await registerAttributionSrc(t, {trigger: {
      event_trigger_data: [{trigger_data: `${value}`}],
    }});

    const payload = await pollEventLevelReports();
    assert_equals(payload.reports.length, 1);
    const report = JSON.parse(payload.reports[0].body);

    const expected = value % cardinality;
    assert_equals(report.trigger_data, `${expected}`);
  }, `Trigger data ${value} is sanitized for source type ${sourceType}.`);

for (let i = 0; i < cardinality * 2; i++) {
  trigger_data_cardinality_promise_test(i);
}
</script>
